Fixed enemy shooting
Enemies now shoot correctly

+++ Project_9_Solution/Source/ModuleParticles.cpp
@@ -123,7 +123,7 @@ Update_Status ModuleParticles::PostUpdate()
 	return Update_Status::UPDATE_CONTINUE;
 }
 
-Particle* ModuleParticles::AddParticle(const Particle& particle, int x, int y, int speedx, int speedy, Collider::Type colliderType, uint delay)
+Particle* ModuleParticles::AddParticle(const Particle& particle, int x, int y, float speedx, float speedy, Collider::Type colliderType, uint delay)
 {
 	Particle* newParticle = nullptr;
